Refactor(core): Fix Dialog circular dependencies
